code: upd timer, res prog added

+++ Iolib.hpp
@@ -1,33 +1,27 @@
 #include <iostream>
 using namespace std;
 
-void Input(int &M, int &L, int &T, int &N, double &Un, double &Uv, int &J){
+void Input(int &M, int &L, double &T, int &N, double &Un, double &Uv, int &J){
     do {
             cout<<"Input sensor channel: ";
             cin>>M;
     } while (M<=0);
-
     do {
             cout<<"Input control channel: ";
             cin>>L;
     } while (L<=0);
-
     do {
             cout<<"Input time interval(in secs): ";
             cin>>T;
     } while (T<=0);
-
     do {
             cout<<"Input number of survey series: ";
             cin>>N;
     } while (N<=0);
-
             cout<<"Input low control action (Un): ";
             cin>>Un;
-
             cout<<"Input high control action (Uv): ";
             cin>>Un;
-
     do {
             cout<<"Input number of surveys in a series: ";
             cin>>J;
